requirements.txt
Django>=4.2
httpx>=0.24.0
python-dotenv